Combine dpc.rs and dpc_queue.rs and move out of architecture specific folder
