+++ xernel/kernel/src/cpu.rs
@@ -1,5 +1,5 @@
 use crate::arch::amd64::apic::APIC;
-use crate::arch::amd64::interrupts::dpc_queue::DpcQueue;
+use crate::dpc::DpcQueue;
 use crate::arch::amd64::{rdmsr, wrmsr, KERNEL_GS_BASE};
 use crate::sched::process::Process;
 use crate::sched::thread::Thread;
